Rearrange Cabal file
I put the top-level sections in alphabetical order. And I put the test
suites in logical order.

1.  HSpec to make sure everything works.
2.  HLint to make sure everything is idiomatic.
3.  Haddock to make sure everything is documented.
4.  HPC to make sure everything is covered.

I know it doesn't matter much because Cabal runs all of them even if any
of them fail. If you're waiting on the results, this is the order you
probably want them in.